requirements.txt
playwright>=1.44.0
httpx>=0.27.0
mapbox-vector-tile>=2.0.3
shapely>=2.0.0
pyproj>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
